+++ scvi/compose/_decorators.py
