+++ scvi/compose/_utils.py
@@ -1,0 +1,7 @@
+import torch
+
+
+def one_hot(index, n_cat):
+    onehot = torch.zeros(index.size(0), n_cat, device=index.device)
+    onehot.scatter_(1, index.type(torch.long), 1)
+    return onehot.type(torch.float32)
