+++ scvi/model/base/_vaemixin.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 from anndata import AnnData
 from torch.distributions import Normal
-from scvi.core._log_likelihood import compute_elbo, compute_reconstruction_error
+from ._log_likelihood import compute_elbo, compute_reconstruction_error
 
 logger = logging.getLogger(__name__)
