+++ scvi/modules/_scanvae.py
@@ -7,11 +7,10 @@ from torch.distributions import kl_divergence as kl
 
 from scvi._compat import Literal
 
-from ._base import Decoder, Encoder
+from scvi.compose import Decoder, Encoder, SCVILoss
+from ._vae import VAE
 from ._classifier import Classifier
 from ._utils import broadcast_labels
-from ._vae import VAE
-from scvi.core.modules._base._base_module import SCVILoss
 from scvi import _CONSTANTS
 
 
@@ -63,7 +62,7 @@ class SCANVAE(VAE):
     use_layer_norm
         Whether to use layer norm in layers
     **kwargs
-        Keyword args for :class:`~scvi.core.modules.VAE`
+        Keyword args for :class:`~scvi.modules.VAE`
     """
 
     def __init__(
