+++ scvi/modules/_vae.py
@@ -9,17 +9,21 @@ from torch.distributions import Normal, Poisson
 from torch.distributions import kl_divergence as kl
 
 from scvi._compat import Literal
-from scvi.core.distributions import (
+from scvi.distributions import (
     NegativeBinomial,
     ZeroInflatedNegativeBinomial,
 )
-from scvi.core.modules._base._base_module import AbstractVAE, SCVILoss
+from scvi.compose import AbstractVAE, SCVILoss
 from scvi import _CONSTANTS
 
 
-from ._base import DecoderSCVI, Encoder, LinearDecoderSCVI
-from ._base.decorators import auto_move_data
-from ._utils import one_hot
+from scvi.compose import (
+    DecoderSCVI,
+    Encoder,
+    LinearDecoderSCVI,
+    auto_move_data,
+    one_hot,
+)
 
 torch.backends.cudnn.benchmark = True
