+++ scvi/modules/_vaec.py
@@ -2,9 +2,9 @@ import torch
 from torch.distributions import Categorical, Normal
 from torch.distributions import kl_divergence as kl
 
-from ._base import DecoderSCVI, Encoder
-from ._classifier import Classifier
-from ._utils import broadcast_labels
+from scvi.compose import DecoderSCVI, Encoder
+from scvi.modules import Classifier
+from scvi.modules._utils import broadcast_labels
 from ._vae import VAE
