+++ tests/core/test_differential.py
@@ -1,6 +1,6 @@
 from scvi.data import synthetic_iid
 from scvi.model import SCVI
-from scvi.core.utils.differential import DifferentialComputation
+from scvi.utils import DifferentialComputation
 from functools import partial
 import numpy as np
 import pytest
